+++ modules/gapi/src/executor/gexecutor.hpp
@@ -8,58 +8,18 @@
 #ifndef OPENCV_GAPI_GEXECUTOR_HPP
 #define OPENCV_GAPI_GEXECUTOR_HPP
 
-#include <memory> // unique_ptr, shared_ptr
-
 #include <utility> // tuple, required by magazine
 #include <unordered_map> // required by magazine
 
-#include <ade/graph.hpp>
-
-#include "backends/common/gbackend.hpp"
+#include "executor/gabstractexecutor.hpp"
 
 namespace cv {
 namespace gimpl {
 
-// Graph-level executor interface.
-//
-// This class specifies API for a "super-executor" which orchestrates
-// the overall Island graph execution.
-//
-// Every Island (subgraph) execution is delegated to a particular
-// backend and is done opaquely to the GExecutor.
-//
-// Inputs to a GExecutor instance are:
-// - GIslandModel - a high-level graph model which may be seen as a
-//   "procedure" to execute.
-//   - GModel - a low-level graph of operations (from which a GIslandModel
-//     is projected)
-// - GComputation runtime arguments - vectors of input/output objects
-//
-// Every GExecutor is responsible for
-// a. Maintaining non-island (intermediate) data objects within graph
-// b. Providing GIslandExecutables with input/output data according to
-//    their protocols
-// c. Triggering execution of GIslandExecutables when task/data dependencies
-//    are met.
-//
-// By default G-API stores all data on host, and cross-Island
-// exchange happens via host buffers (and CV data objects).
-//
-// Today's exchange data objects are:
-// - cv::Mat               - for image buffers
-// - cv::Scalar            - for single values (with up to four components inside)
-// - cv::detail::VectorRef - an untyped wrapper over std::vector<T>
-//
-
-class GExecutor
+class GExecutor final: public GAbstractExecutor
 {
 protected:
     Mag m_res;
-    std::unique_ptr<ade::Graph> m_orig_graph;
-    std::shared_ptr<ade::Graph> m_island_graph;
-
-    cv::gimpl::GModel::Graph       m_gm;  // FIXME: make const?
-    cv::gimpl::GIslandModel::Graph m_gim; // FIXME: make const?
 
     // FIXME: Naive executor details are here for now
     // but then it should be moved to another place
@@ -85,14 +45,12 @@ protected:
 
 public:
     explicit GExecutor(std::unique_ptr<ade::Graph> &&g_model);
-    void run(cv::gimpl::GRuntimeArgs &&args);
+    void run(cv::gimpl::GRuntimeArgs &&args) override;
 
-    bool canReshape() const;
-    void reshape(const GMetaArgs& inMetas, const GCompileArgs& args);
+    bool canReshape() const override;
+    void reshape(const GMetaArgs& inMetas, const GCompileArgs& args) override;
 
-    void prepareForNewStream();
-
-    const GModel::Graph& model() const; // FIXME: make it ConstGraph?
+    void prepareForNewStream() override;
 };
 
 } // namespace gimpl
